Add option to disable chunked downloads
Some browsers have issues with chunked downloads. This commit adds
the ability to request non chunked downloads. This comes at the cost
of creating a temporary file, as logs are being kept inside a time
series database, and need to be dumped to a temporary file before
sending it off to the client. This also means we add a time penalty
when downloading the log file.